moved shared components into its own folder
