+++ package.json
@@ -12,7 +12,6 @@
     "@supabase/ssr": "^0.5.0",
     "@supabase/supabase-js": "^2.45.2",
     "autoprefixer": "^10.4.20",
-    "eslint": "^8.57.1",
     "next": "14.2.6",
     "next-cloudinary": "^6.12.0",
     "postcss": "^8.4.47",
@@ -22,10 +21,15 @@
     "tailwindcss": "^3.4.12"
   },
   "devDependencies": {
+    "@eslint/js": "^9.24.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "eslint": "^8.57.1",
     "eslint-config-next": "14.2.6",
-    "typescript": "^5"
+    "eslint-plugin-react": "^7.37.5",
+    "globals": "^16.0.0",
+    "typescript": "^5",
+    "typescript-eslint": "^8.30.0"
   }
 }
