+++ src/app/about/changelog/page.jsx
@@ -1,4 +1,4 @@
-import Header from '@/app/homepage/header';
+import Header from '@/app/_components/header'
 
 export default function Changelog() {
     return (
