+++ src/app/page.jsx
@@ -1,12 +1,12 @@
 import ReleaseTable from './releases/release-table.js';
 import LoginForm from './login/login-form.js';
-import Header from './homepage/header.js'
+import Header from '@/app/_components/header'
 import './globals.css'
 import 'tailwindcss/tailwind.css'
-import Disclaimer from './homepage/disclaimer.js'
-import Footer from './homepage/footer.js'
-import SearchComponent from './search/search-component.js';
-import SearchSideBar from './search/search-sidebar.js'
+import HowToUse from './_components/how-to-use.js'
+import Footer from '@/app/_components/footer'
+import SearchComponent from './_components/_search/search-by-name.js';
+import CategoryFilter from './_components/_search/category-filter.js';
 
 export default function Home({ params, searchParams }) {
   return (
@@ -15,11 +15,11 @@ export default function Home({ params, searchParams }) {
       <div>
         <div className="grid grid-cols-[20%_80%] px-5 pt-5">
           <SearchComponent params={""} />
-          <Disclaimer className="px-5"/>
+          <HowToUse className="px-5"/>
         </div>
       </div>
       <div className="grid grid-cols-[20%_80%] p-5">
-        <SearchSideBar />
+        <CategoryFilter />
         <ReleaseTable params={params} searchParams={searchParams} />
       </div>
       <Footer />
