+++ src/app/profile/user-analytics.js
@@ -17,6 +17,7 @@ async function calculateTotal(money) {
 export default async function UserAnalytics() {
     const supabase = createClient();
     const {data: { user }} = await supabase.auth.getUser();
+    // fetch all user data. 
     const {data: userData, error: inventoryError} = await supabase.from('user_data').select().eq('user_id', user.id);
 
     // convert user data to array of yen spent
