+++ tsconfig.json
@@ -21,6 +21,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/app/page.tsx", "src/app/releases/release_table.js", "src/app/profile/page.js", "src/app/releases/[id]/page.js", "src/app/releases/[id]/page.jsx", "src/app/login/page.jsx", "src/app/search/search.js", "src/app/search/search-bar.js", "src/app/releases/release-table.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/app/page.tsx", "src/app/releases/release_table.js", "src/app/profile/page.js", "src/app/releases/[id]/page.js", "src/app/releases/[id]/page.jsx", "src/app/login/page.jsx", "src/app/_components/_search/search.js", "src/app/_components/_search/search-bar.js", "src/app/releases/release-table.js"],
   "exclude": ["node_modules"]
 }
